Split the Robot class out to a separate file
